feat: subdocument operations

+++ api/controllers/campground.controller.js
@@ -1,32 +1,57 @@
 const mongoose = require("mongoose");
 const Campground = mongoose.model(process.env.MODEL_NAME_CAMPGROUND);
 require("dotenv").config();
+const { sendResponse, checkObjectIdValid } = require("../helper");
 
 const getAll = function (req, res) {
     let offset = parseInt(process.env.DEFAULT_FIND_OFFSET, process.env.NUMBER_BASE);
     let count = parseInt(process.env.DEFAULT_FIND_COUNT, process.env.NUMBER_BASE);
+    
+    if (req.query.offset) {
+        if(!isNaN(req.query.offset)){
+            offset =  parseInt(req.query.offset, process.env.NUMBER_BASE);
+        } else {
+            sendResponse(res, process.env.STATUS_CODE_BAD_REQUEST, { message: process.env.MESSAGE_QUERY_PARAM_NOT_VALID });
+            return;
+        }
+    }
+    if (req.query.count) {
+        if(!isNaN(req.query.count)){
+            count =  parseInt(req.query.count, process.env.NUMBER_BASE);
+            
+            if(count > parseInt(process.env.DEFAULT_MAX_COUNT, process.env.NUMBER_BASE)) {
+                sendResponse(res, process.env.STATUS_CODE_BAD_REQUEST, { message: process.env.MESSAGE_MAX_COUNT_VALUE_ERROR });
+                return;
+            }
+        } else {
+            sendResponse(res, process.env.STATUS_CODE_BAD_REQUEST, { message: process.env.MESSAGE_QUERY_PARAM_NOT_VALID });
+            return;
+        }
+    }
+
     Campground.find().skip(offset).limit(count).exec(function (err, campgrounds) {
         if (err) {
-            res.status(parseInt(process.env.STATUS_CODE_ERROR)).send({ message: process.env.MESSAGE_ERROR });
+            sendResponse(res, process.env.STATUS_CODE_ERROR, { message: process.env.MESSAGE_ERROR })
         } else {
             if (!campgrounds) {
-                res.status(parseInt(process.env.STATUS_CODE_NOT_FOUND)).json({ message: process.env.MESSAGE_NOT_FOUND });
+                sendResponse(res, process.env.STATUS_CODE_NOT_FOUND, { message: process.env.MESSAGE_NOT_FOUND })
             } else {
-                console.log("Found campgrounds", campgrounds);
-                res.status(parseInt(process.env.STATUS_CODE_OK)).json(campgrounds);
+                // console.log("Found campgrounds", campgrounds);
+                sendResponse(res, process.env.STATUS_CODE_OK, campgrounds);
             }
         }
     });
 };
 
-const addOne = function (req, res) {
+const addOneOrMore = function (req, res) {
     const newCampgroundData = req.body;
     Campground.create(newCampgroundData, function (err, campground) {
         if (err) {
-            res.status(parseInt(process.env.STATUS_CODE_ERROR)).send({ message: process.env.MESSAGE_ERROR });
+            console.log("Internal error", err)
+            sendResponse(res, process.env.STATUS_CODE_ERROR, { message: process.env.MESSAGE_ERROR })
         } else {
             console.log("Campground created", campground);
-            res.status(parseInt(process.env.STATUS_CODE_OK)).json(campground);
+            sendResponse(res, process.env.STATUS_CODE_CREATED, { message: process.env.MESSAGE_NEW_ITEM_OR_ITEMS_CREATED });
         }
     })
 };
@@ -35,13 +60,13 @@ const getOne = function (req, res) {
     const campgroundId = req.params.campgroundId;
     Campground.findById(campgroundId).exec(function (err, campground) {
         if (err) {
-            res.status(parseInt(process.env.STATUS_CODE_ERROR)).send({ message: process.env.MESSAGE_ERROR });
+            sendResponse(res, process.env.STATUS_CODE_ERROR, { message: process.env.MESSAGE_ERROR })
         } else {
             if (!campground) {
-                res.status(parseInt(process.env.STATUS_CODE_NOT_FOUND)).json({ message: process.env.MESSAGE_NOT_FOUND });
+                sendResponse(res, process.env.STATUS_CODE_NOT_FOUND, { message: process.env.MESSAGE_NOT_FOUND })
             } else {
                 console.log("Found campground by id", campground);
-                res.status(parseInt(process.env.STATUS_CODE_OK)).json(campground);
+                sendResponse(res, process.env.STATUS_CODE_OK, campground);
             }
         }
     })
@@ -49,15 +74,20 @@ const getOne = function (req, res) {
 
 const deleteOne = function (req, res) {
     const campgroundId = req.params.campgroundId;
+
+    if (checkObjectIdValid(res, campgroundId)) {
+        return;
+    }
+
     Campground.findByIdAndDelete(campgroundId).exec(function (err, campground) {
         if (err) {
-            res.status(parseInt(process.env.STATUS_CODE_ERROR)).send({ message: process.env.MESSAGE_ERROR });
+            sendResponse(res, process.env.STATUS_CODE_ERROR, { message: process.env.MESSAGE_ERROR })
         } else {
             if (!campground) {
-                res.status(parseInt(process.env.STATUS_CODE_NOT_FOUND)).json({ message: process.env.MESSAGE_NOT_FOUND });
+                sendResponse(res, process.env.STATUS_CODE_NOT_FOUND, { message: process.env.MESSAGE_NOT_FOUND })
             } else {
                 console.log("Deleted campground", campground);
-                res.status(parseInt(process.env.STATUS_CODE_OK)).json(campground);
+                sendResponse(res, process.env.STATUS_CODE_OK, { message: process.env.MESSAGE_DELETE_ITEM });
             }
         }
     })
@@ -65,19 +95,29 @@ const deleteOne = function (req, res) {
 
 const fullUpdateOne = function (req, res) {
     const campgroundId = req.params.campgroundId;
+    
+    if (checkObjectIdValid(res, campgroundId)) {
+        return;
+    }
+
     Campground.findById(campgroundId).exec(function (err, campground) {
         if (err) {
-            res.status(parseInt(process.env.STATUS_CODE_ERROR)).send({ message: process.env.MESSAGE_ERROR });
+            sendResponse(res, process.env.STATUS_CODE_ERROR, { message: process.env.MESSAGE_ERROR })
         } else {
             if (!campground) {
-                res.status(parseInt(process.env.STATUS_CODE_NOT_FOUND)).json({ message: process.env.MESSAGE_NOT_FOUND });
+                sendResponse(res, process.env.STATUS_CODE_NOT_FOUND, { message: process.env.MESSAGE_NOT_FOUND })
             } else {
                 campground.name = req.body.name;
-                campground.city = req.body.city;
                 campground.state = req.body.state;
-                campground.save(function () {
-                    console.log("Deleted campground", campground);
-                    res.status(parseInt(process.env.STATUS_CODE_OK)).json(campground);
+                campground.park = req.body.park;
+                campground.species = req.body.species ? req.body.species : [];
+                campground.save(function (err) {
+                    if (err) {
+                        sendResponse(res, process.env.STATUS_CODE_ERROR, { message: process.env.MESSAGE_REQUIRED_FIELDS_MISSING })
+                    } else {
+                        console.log("Full update one campground", campground);
+                        sendResponse(res, process.env.STATUS_CODE_OK, { message: process.env.MESSAGE_PARTIAL_UPDATE });
+                    }
                 });
             }
         }
@@ -86,25 +126,37 @@ const fullUpdateOne = function (req, res) {
 
 const partialUpdateOne = function (req, res) {
     const campgroundId = req.params.campgroundId;
+    
+    if (checkObjectIdValid(res, campgroundId)) {
+        return;
+    }
+
     Campground.findById(campgroundId).exec(function (err, campground) {
         if (err) {
-            res.status(parseInt(process.env.STATUS_CODE_ERROR)).send({ message: process.env.MESSAGE_ERROR });
+            sendResponse(res, process.env.STATUS_CODE_ERROR, { message: process.env.MESSAGE_ERROR })
         } else {
             if (!campground) {
-                res.status(parseInt(process.env.STATUS_CODE_NOT_FOUND)).json({ message: process.env.MESSAGE_NOT_FOUND });
+                sendResponse(res, process.env.STATUS_CODE_NOT_FOUND, { message: process.env.MESSAGE_NOT_FOUND })
             } else {
                 if (req.body.name) {
                     campground.name = req.body.name;
                 }
-                if (req.body.city) {
-                    campground.city = req.body.city;
-                }
                 if (req.body.state) {
                     campground.state = req.body.state;
                 }
-                campground.save(function () {
-                    console.log("Campground partially updated", campground);
-                    res.status(parseInt(process.env.STATUS_CODE_OK)).json(campground);
+                if (req.body.park) {
+                    campground.park = req.body.park;
+                }
+                if (req.body.species) {
+                    campground.species = req.body.species;
+                }
+                campground.save(function (err) {
+                    if (err) {
+                        sendResponse(res, process.env.STATUS_CODE_ERROR, { message: process.env.MESSAGE_REQUIRED_FIELDS_MISSING })
+                    } else {
+                        console.log("Campground partially updated", campground);
+                        sendResponse(res, process.env.STATUS_CODE_OK, { message: process.env.MESSAGE_PARTIAL_UPDATE });
+                    }
                 });
             }
         }
@@ -113,7 +165,7 @@ const partialUpdateOne = function (req, res) {
 
 module.exports = {
     getAll,
-    addOne,
+    addOneOrMore,
     getOne,
     deleteOne,
     fullUpdateOne,
